Move sequential members to another part
